Quests Code Review
