+++ unity-client/Assets/Scripts/MainScripts/DCL/Bridges/Bridges.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 2356216790431665206}
   - component: {fileID: 7384471490513260130}
   - component: {fileID: 1314399029881169376}
+  - component: {fileID: 3653805652683081651}
   m_Layer: 0
   m_Name: Bridges
   m_TagString: Untagged
@@ -80,5 +81,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: cf40fa55fbd90154dbf7ff397c0eca1a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &3653805652683081651
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6401575379531892522}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1735a895abd8bc1469f4361f6bc37940, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
