+++ unity-client/Assets/Scripts/MainScripts/DCL/Bridges/QuestsBridge.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0581875e24a92d641ae86d8a8f41d1ab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
