+++ unity-client/Assets/Scripts/MainScripts/DCL/Bridges/QuestsBridge/QuestsBridge.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "QuestsBridge",
+    "references": [
+        "GUID:9b06d63e1a13bc34c995683595d8d849",
+        "GUID:81a6b1347b86bfe4aa02ed089ff533c1",
+        "GUID:b1087c5731ff68448a0a9c625bb7e52d"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
