+++ unity-client/Assets/Scripts/MainScripts/DCL/Bridges/QuestsBridge/QuestsBridge.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ca1ff26951d572045ae4605e8cb7322d
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
