+++ unity-client/Assets/Scripts/MainScripts/DCL/Bridges/QuestsBridge/QuestsBridge.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1735a895abd8bc1469f4361f6bc37940
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
